+++ docs/en/CHANGELOG.md
@@ -1,13 +1,14 @@
 ## 0.6.3 (2025-08-11)
-- Add locate game window
 - Rework -d/--data feature
 - Add card and gift to result table
-- FB app now reset friction brake at startup
+- FB and app now reset friction brake at startup
 - Faster FB Reset
 - Now FB app use FRICTION_BRAKE.SENSITIVITY correctly
 - Now Bot app use BOT.FRICTION_BRAKE.INITIAL correctly
 - Fix pause bug
 - Optimize time required to pause
+- Now reset fb on quitting
+
 
 ## 0.6.3 (2025-08-11)